requirements.txt
pip~=24.2
pyserial~=3.5
websockets~=15.0.1